Fix #654 - warn about MixedArgument when inferredType property does not exist
And handle str_replace more leniently when args unknown

+++ src/Psalm/Codebase/Reflection.php
@@ -155,7 +155,7 @@ class Reflection
         }
 
         $reflection_methods = $reflected_class->getMethods(
-            \ReflectionMethod::IS_PUBLIC | \ReflectionMethod::IS_PROTECTED
+            (int) (\ReflectionMethod::IS_PUBLIC | \ReflectionMethod::IS_PROTECTED)
         );
 
         if ($class_name_lower === 'generator') {
